rename module: instant_substraction to time_substraction
